backend/pyproject.toml
[project]
name = "pichunter-backend"
version = "0.1.0"
description = "Backend service for the Pichunter Figma plugin"
requires-python = ">=3.10"

# 项目依赖
# fastapi: 高性能 Web API 框架
# uvicorn[standard]: ASGI 服务器，支持异步和热重载
# python-multipart: 处理表单和文件上传
# pillow: 图片处理库
# httpx: 异步 HTTP 客户端
dependencies = [
  "fastapi==0.115.5",           # 核心框架：类似 UIKit/SwiftUI，负责构建 App 的骨架和逻辑
  "uvicorn[standard]==0.32.0",  # 服务器：类似 iOS 系统 (SpringBoard)，负责监听网络端口并唤醒 App (FastAPI)
  "python-multipart==0.0.9",    # 数据解析：类似 JSONDecoder，但专门用来解析文件上传 (Multipart) 数据
  "pillow==11.0.0",             # 图片处理：类似 CoreGraphics / UIImage，用来处理图片数据               # 网络请求：类似 URLSession，让后端代码可以发起网络请求 (比如调用 OpenAI)
  "httpx==0.27.2",
  "google-generativeai>=0.8.5",
]

[tool.uv]
# uv 工具配置，package=false 表示不打包项目
package = false

[tool.uv.scripts]
# uv 自定义脚本
# dev: 启动开发服务器，使用 uvicorn 运行 FastAPI 应用，支持热重载，监听 8000 端口
dev = "uvicorn app.main:app --reload --port 8000"
